bug import product LDLC
Former-commit-id: 17fe7218d339495754b62240ad010409f7bbdd00

+++ htdocs/bimpdatasync/classes/process_overrides/BDS_ImportsAppleProcess.php
@@ -426,7 +426,7 @@ class BDS_ImportsAppleProcess extends BDSImportProcess
                                         break;
 
                                     case 'price':
-                                        $value = (float) str_replace(' ', '', $value);
+                                        $value = str_replace(' ', '', $value);
                                         $value = (float) str_replace(',', '.', $value);
                                         break;
                                 }
